feat: full is done

+++ backend/src/books/books.service.ts
@@ -4,10 +4,14 @@ import { Model } from 'mongoose';
 import { Book } from './schemas/books.schema';
 import { CreateBookDto } from './dto/createBook.dto';
 import { UpdateBookDto } from './dto/updateBook.dto';
+import { Author } from 'src/authors/schemas/authors.schema';
 
 @Injectable()
 export class BooksService {
-  constructor(@InjectModel(Book.name) private readonly bookModel: Model<Book>) {}
+  constructor(
+    @InjectModel(Book.name) private readonly bookModel: Model<Book>,
+    @InjectModel(Author.name) private readonly authorModel: Model<Author>,
+  ) {}
 
   async create(createBookDto: CreateBookDto): Promise<Book> {
     const book = new this.bookModel(createBookDto);
@@ -15,11 +19,11 @@ export class BooksService {
   }
 
   async findAll(): Promise<Book[]> {
-    return this.bookModel.find().populate('authorIds').exec();
+    return this.bookModel.find().populate('authorId').exec();
   }
 
   async findOne(id: string): Promise<Book> {
-    const book = await this.bookModel.findById(id).populate('authorIds').exec();
+    const book = await this.bookModel.findById(id).populate('authorId').exec();
     if (!book) {
       throw new NotFoundException(`Book with ID "${id}" not found`);
     }
@@ -29,7 +33,7 @@ export class BooksService {
   async update(id: string, updateBookDto: UpdateBookDto): Promise<Book> {
     const updatedBook = await this.bookModel
       .findByIdAndUpdate(id, updateBookDto, { new: true })
-      .populate('authorIds')
+      .populate('authorId')
       .exec();
     if (!updatedBook) {
       throw new NotFoundException(`Book with ID "${id}" not found`);
@@ -38,9 +42,23 @@ export class BooksService {
   }
 
   async delete(id: string): Promise<void> {
+    // Найти книгу, чтобы получить ID автора
+    const book = await this.bookModel.findById(id).exec();
+    if (!book) {
+      throw new NotFoundException(`Book with ID "${id}" not found`);
+    }
+  
+    // Удалить книгу
     const result = await this.bookModel.findByIdAndDelete(id).exec();
     if (!result) {
       throw new NotFoundException(`Book with ID "${id}" not found`);
     }
+  
+    // Удалить ID книги из списка книг автора
+    await this.authorModel.updateOne(
+      { _id: book.authorId },
+      { $pull: { books: id } }
+    );
   }
+  
 }
